devkit-utils: speed up call to sync
We don't need to wait for all filesystems to be synced before changing
the client. We just need to ensure that writes to the devkit working
folder are flushed.

This helps in cases where steam is downloading a game in the background
and there are a lot of pending writes to flush.